Fix to work with YARD v0.9.25+
From v0.9.25, YARD changed to use RedCarpert's method of trying to
create GitHub-style anchor links. RedCarpet does NOT match GitHub's
algorithm exactly, so it all got messed up. In this commit, the
code is changed to grab the new 'id="..."' field from '<h\d+...'
tags and use that as the YARD ID. I tried recreating RedCarpert's
C code (rndr_header_anchor() in 'ext/redcarpet/html.c') but this
failed miserably, so resorted to just this. All that matters is
that it works!